+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Full changelog for the costs to expect REST API.
 * Added the ability to POST a request error to the API.
 * Added request/error-log route.
 * Added request/log route.
+* Reworked pagination utility class.
+* Modified HEADER links for pagination.
 * Removed all code referencing PATCH and update, not ready to implement yet and may modify design.
 * Minor refactoring, order of method params etc.